pasantias/requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.10
django-disqus==0.5
django-postgrespool==0.3.0
django-toolbelt==0.0.1
facebook-sdk==0.4.0
funcsigs==0.4
gunicorn==19.3.0
httplib2==0.9.1
mock==1.3.0
oauth2==1.9.0.post1
oauthlib==1.0.3
pbr==1.8.0
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.12
reportlab==3.2.0
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
SQLAlchemy==1.0.8
static3==0.6.1
Unipath==1.1
virtualenv==13.1.2
virtualenvwrapper-win==1.2.1
wheel==0.24.0
